move duplicate function to common functions
